+++ prerender/routes.js
@@ -1,0 +1,5 @@
+import { getBlogRoutes } from './blog-routes.js';
+
+export function getPrerenderRoutes() {
+  return ['/', ...getBlogRoutes()];
+}